fb: Support 64-bit EFI framebuffer bases

+++ stage23/lib/fb.h
@@ -11,7 +11,7 @@ struct resolution {
 
 struct fb_info {
     uint8_t  memory_model;
-    uint32_t framebuffer_addr;
+    uint64_t framebuffer_addr;
     uint16_t framebuffer_pitch;
     uint16_t framebuffer_width;
     uint16_t framebuffer_height;
